修改了一些文件
AI、UCT、MONTECALO文件更新
BOARD、MONTECALObug修正

+++ interface.cpp
@@ -50,7 +50,7 @@ static struct gtp_command commands[] = {
 
 
 extern int board_size;
-extern float komi;
+extern double komi;
 
 int
 main(int argc, char **argv)
